# docs/v1_swagger.yaml
basePath: /api/v1
definitions:
  biz.Biz:
    properties:
      id:
        example: id123zqqeeadg24qasd
        type: string
      info:
        example: input info here
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
    type: object
  errdef.Err:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact:
    email: support@sinlov.cn
    name: API Support
    url: http://github.com/
  description: This is a sample server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://github.com/
  title: gin-api-swagger-temple
  version: v1.x.x
paths:
  /biz/form:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: warning api in prod will hide, abs remote api for dev
      parameters:
      - example: id123zqqeeadg24qasd
        in: formData
        name: id
        type: string
      - example: input info here
        in: formData
        name: info
        type: string
      - example: 10
        in: formData
        name: limit
        type: integer
      - example: 0
        in: formData
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: value in model.Biz
          schema:
            $ref: '#/definitions/biz.Biz'
        "400":
          description: error at errdef.Err
          schema:
            $ref: '#/definitions/errdef.Err'
      summary: 'api demo form with: x-www-form-urlencoded'
      tags:
      - biz
  /biz/json:
    get:
      consumes:
      - application/json
      description: warning api in prod will hide, abs remote api for dev
      produces:
      - application/json
      responses:
        "200":
          description: value in biz.Biz
          schema:
            $ref: '#/definitions/biz.Biz'
        "500":
          description: ""
      summary: api demo json
      tags:
      - biz
  /biz/modelBiz:
    post:
      consumes:
      - application/json
      description: warning api in prod will hide, abs remote api for dev
      parameters:
      - description: body biz.Biz for post
        in: body
        name: biz
        required: true
        schema:
          $ref: '#/definitions/biz.Biz'
      produces:
      - application/json
      responses:
        "200":
          description: value in biz.Biz
          schema:
            $ref: '#/definitions/biz.Biz'
        "400":
          description: error at errdef.Err
          schema:
            $ref: '#/definitions/errdef.Err'
      summary: api json struct biz.Biz
      tags:
      - biz
  /biz/modelBizQuery:
    post:
      consumes:
      - application/json
      description: warning api in prod will hide, abs remote api for dev
      parameters:
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: body biz.Biz for post
        in: body
        name: biz
        required: true
        schema:
          $ref: '#/definitions/biz.Biz'
      produces:
      - application/json
      responses:
        "200":
          description: value in biz.Biz
          schema:
            $ref: '#/definitions/biz.Biz'
        "400":
          description: error at errdef.Err
          schema:
            $ref: '#/definitions/errdef.Err'
      summary: api post query with json struct biz.Biz
      tags:
      - biz
  /biz/path/{some_id}:
    get:
      consumes:
      - multipart/form-data
      description: warning api in prod will hide, abs remote api for dev
      parameters:
      - description: some id to show
        in: path
        name: some_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: value in biz.Biz
          schema:
            $ref: '#/definitions/biz.Biz'
        "400":
          description: error at errdef.Err
          schema:
            $ref: '#/definitions/errdef.Err'
      summary: api path demo for route path
      tags:
      - biz
  /biz/query/:
    get:
      consumes:
      - application/json
      description: warning api in prod will hide, abs remote api for dev
      parameters:
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: value in biz.Biz
          schema:
            $ref: '#/definitions/biz.Biz'
        "400":
          description: error at errdef.Err
          schema:
            $ref: '#/definitions/errdef.Err'
      summary: api demo for query.
      tags:
      - biz
  /biz/string:
    get:
      consumes:
      - application/json
      description: get string of this api. warning api in prod will hide, abs remote
        api for dev
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "500":
          description: ""
      summary: sample demo string
      tags:
      - biz
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload file. warning api in prod will hide, abs remote api for
        dev
      operationId: file.upload
      parameters:
      - description: this is a test file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/errdef.Err'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/errdef.Err'
      summary: Upload file for count file size
      tags:
      - biz
securityDefinitions:
  BasicAuth:
    type: basic
  WithToken:
    description: Please set the token of the API, note that it starts with "Bearer"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"